+++ python/src/models/deep/competitive/__init__.py
@@ -1,3 +1,10 @@
+"""
+Competitive Learning Networks.
+
+Implementations of competitive learning algorithms including:
+- Self-Organizing Maps (SOM) / Kohonen Maps
+- Learning Vector Quantization (LVQ)
+"""
 from .som import KohonenMap
 from .lvq import LVQ